+++ tests/conftest.py
@@ -33,14 +33,22 @@ def db():
         cur = conn.cursor()
 
         try:
+            cur.execute("CREATE TABLE collection (id int, source_id text, transform_from_collection_id int)")
             cur.execute("CREATE TABLE release (id int, collection_id int, ocid text, data_id int)")
             cur.execute("CREATE TABLE record (id int, collection_id int, ocid text, data_id int)")
             cur.execute("CREATE TABLE data (id int, data jsonb)")
+
+            cur.execute("INSERT INTO collection VALUES (1, 'scotland', NULL)")
+            cur.execute("INSERT INTO collection VALUES (2, 'paraguay_dncp_records', NULL)")
+            cur.execute("INSERT INTO collection VALUES (3, 'paraguay_dncp_releases', NULL)")
+            cur.execute("INSERT INTO collection VALUES (4, 'paraguay_dncp_releases', 3)")
+            cur.execute("INSERT INTO collection VALUES (5, 'paraguay_dncp_releases', 4)")
             cur.execute("INSERT INTO release VALUES (1, 1, 'ocds-213czf-1', 1)")
             cur.execute("INSERT INTO record VALUES (1, 1, 'ocds-213czf-2', 2)")
             cur.execute("""INSERT INTO data VALUES (1, '{"ocid":"ocds-213czf-1"}'::jsonb)""")
             cur.execute("""INSERT INTO data VALUES (2, '{"ocid":"ocds-213czf-2","""
                         """"releases":[{"ocid":"ocds-213czf-2"}]}'::jsonb)""")
+
             conn.commit()
 
             yield